split color into simple (copy) and complex (clone)
